Splitting the barlow_twins.py file to have the LinearClassifierModule in separate linear.py file
